+++ psi4/src/psi4/dfmp2/corr_grad.h
@@ -50,7 +50,7 @@ protected:
     /// Bench flag, defaults to 0
     int bench_;
     /// Memory available, in doubles, defaults to 256 MB (32 M doubles)
-    unsigned long int memory_;
+    size_t memory_;
     /// Number of OpenMP threads (defaults to 1 in no OpenMP, Process::environment.get_n_threads() otherwise)
     int nthreads_;
     /// Integral cutoff (defaults to 0.0)
@@ -123,7 +123,7 @@ public:
      * integral generation objects typically ignore this)
      * @param memory maximum number of doubles to allocate
      */
-    void set_memory(unsigned long int memory) { memory_ = memory; }
+    void set_memory(size_t memory) { memory_ = memory; }
     /**
      * Maximum number of OpenMP threads to use. It may be necessary
      * to clamp this to some value smaller than the total number of
@@ -171,8 +171,8 @@ protected:
     void build_AB_x_terms();
     void build_Amn_x_terms();
 
-    void fitting_helper(SharedMatrix J, unsigned int file, const std::string& label, unsigned long int naux, unsigned long int nij, unsigned long int memory);
-    void UV_helper(SharedMatrix V, double c, unsigned int file, const std::string& label, unsigned long int naux, unsigned long int nij, unsigned long int memory);
+    void fitting_helper(SharedMatrix J, unsigned int file, const std::string& label, size_t naux, size_t nij, size_t memory);
+    void UV_helper(SharedMatrix V, double c, unsigned int file, const std::string& label, size_t naux, size_t nij, size_t memory);
 
     /// File number for Alpha (Q|mn) tensor
     unsigned int unit_a_;
