+++ psi4/src/psi4/lib3index/cholesky.cc
@@ -50,7 +50,7 @@
 
 namespace psi {
 
-Cholesky::Cholesky(double delta, unsigned long int memory)
+Cholesky::Cholesky(double delta, size_t memory)
     : delta_(delta), memory_(memory), Q_(0)
 {
 }
@@ -148,7 +148,7 @@ void Cholesky::choleskify()
     }
 }
 
-CholeskyMatrix::CholeskyMatrix(SharedMatrix A, double delta, unsigned long int memory) :
+CholeskyMatrix::CholeskyMatrix(SharedMatrix A, double delta, size_t memory) :
     A_(A), Cholesky(delta, memory)
 {
     if (A_->nirrep() != 1)
@@ -177,7 +177,7 @@ void CholeskyMatrix::compute_row(int row, double* target)
 }
 
 CholeskyERI::CholeskyERI(std::shared_ptr<TwoBodyAOInt> integral, double schwarz,
-    double delta, unsigned long int memory) :
+    double delta, size_t memory) :
     integral_(integral), schwarz_(schwarz), Cholesky(delta, memory)
 {
     basisset_ = integral_->basis();
@@ -253,7 +253,7 @@ CholeskyMP2::CholeskyMP2(SharedMatrix Qia,
     std::shared_ptr<Vector> eps_aocc,
     std::shared_ptr<Vector> eps_avir,
     bool symmetric,
-    double delta, unsigned long int memory) :
+    double delta, size_t memory) :
     Qia_(Qia), eps_aocc_(eps_aocc), eps_avir_(eps_avir),
     symmetric_(symmetric), Cholesky(delta, memory)
 {
@@ -306,7 +306,7 @@ void CholeskyMP2::compute_row(int row, double* target)
 CholeskyDelta::CholeskyDelta(
     std::shared_ptr<Vector> eps_aocc,
     std::shared_ptr<Vector> eps_avir,
-    double delta, unsigned long int memory) :
+    double delta, size_t memory) :
     eps_aocc_(eps_aocc), eps_avir_(eps_avir),
     Cholesky(delta, memory)
 {
@@ -352,7 +352,7 @@ void CholeskyDelta::compute_row(int row, double* target)
 
 CholeskyLocal::CholeskyLocal(
     SharedMatrix C,
-    double delta, unsigned long int memory) :
+    double delta, size_t memory) :
     C_(C), Cholesky(delta, memory)
 {
 }
