+++ psi4/src/psi4/libpsio/toclen.cc
@@ -106,7 +106,7 @@ void PSIO::wt_toclen(unsigned int unit, ULI len) {
   }
 }
 
-unsigned long int psio_rd_toclen(unsigned int unit) {
+size_t psio_rd_toclen(unsigned int unit) {
   return _default_psio_lib_->rd_toclen(unit);
 }
